requirements.txt
# Django & Core Dependencies
Django==5.1.4
asgiref==3.8.1
sqlparse==0.5.3

# Authentication & Security
django-allauth==65.3.0
PyJWT==2.10.1
python-jose==3.3.0
oauthlib==3.2.2
cryptography==44.0.0
pyotp==2.9.0
qrcode==8.0

# Environment & Configuration
django-environ==0.11.2
python-dotenv==1.0.1

# Frontend & UI
django-htmx==1.21.0
django-tailwind==3.8.0
Pillow==11.0.0

# API Integration
plaid-python==28.0.0
requests==2.32.3
urllib3==2.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10

# Testing
pytest==8.3.4
pytest-django==4.9.0
pluggy==1.5.0
iniconfig==2.0.0
packaging==24.2

# Utilities
python-dateutil==2.9.0.post0
six==1.17.0
pyasn1==0.6.1
rsa==4.9
ecdsa==0.19.0

# Development Tools
rich==13.9.4
Pygments==2.18.0
markdown-it-py==3.0.0
mdurl==0.1.2
